requirements.txt
fastapi==0.110.1
fastapi-sso==0.14.2
fastapi-utilities==0.2.1a0
konlpy==0.6.0
pydantic==2.6.4
pydantic_core==2.16.3
slowapi==0.1.9
SQLAlchemy==2.0.29
starlette==0.37.2
torch==2.2.2
transformers==4.39.3
uvicorn==0.29.0
wheel==0.43.0
python-jose==3.3.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.0.3
alembic==1.13.1
alembic-postgresql-enum==1.2.0